% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Country Population}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{A string that represents the name of the country}
}
\value{
A ggplot object of population over time of the input country
}
\description{
Country population is a function that takes as an input the string name of
a country and returns the graph of estimated population over time. The data
is obtained using the World Population data that is found in this
package.
}
\examples{
CountryPopulation("China")
CountryPopulation("Tuvalu")
}
