% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup attendance data}
\format{
A data frame with 22 observations with 4 columns.
\describe{
   \item{Totalattendance}{The attendance for each match held at that world cup.}
   \item{Matches}{The number of matches held for that world cup.}
   \item{Averageattendance}{The averaged attendance for each match.}
   \item{WorldCup}{A string containg the location and year the world cup was held. }

}
}
\usage{
World_Cup
}
\description{
This is a data table with data concerning the average attendance at every
world cup, the year and where it is held.
}
\keyword{datasets}
