% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Data}
\format{
A data frame with 235 observations with 72 columns.
\describe{
   \item{Country}{The name of the country as recognized by the UN}

}
}
\usage{
WorldPopulation
}
\description{
A data table containing the population data of each country in the United
Nations from 1950 to 2020. The data was collected by the United Nations
and was obtained from the final module in Canvas
}
\keyword{datasets}
